pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-maker-auto"
version = "1.0.0"
description = "Combine a static image + audio into a YouTube-optimized MP4 video."
requires-python = ">=3.10"
dependencies = [
    "Pillow>=11.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.0",
]

[project.scripts]
video-maker = "video_maker.cli:main"

[tool.setuptools.packages.find]
include = ["video_maker*"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that require a real ffmpeg binary (deselect with '-m \"not integration\"')",
]
addopts = "-m 'not integration'"
